Moved operators functions from Expression to our reflectmath.go
