+++ ast/Expression.go
@@ -4,9 +4,10 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"github.com/google/uuid"
 	"reflect"
-	"time"
+
+	"github.com/google/uuid"
+	"github.com/hyperjumptech/grule-rule-engine/pkg"
 )
 
 const (
@@ -203,35 +204,35 @@ func (e *Expression) Evaluate() (reflect.Value, error) {
 
 		switch e.Operator {
 		case OpMul:
-			val, opErr = EvaluateMultiplication(lval, rval)
+			val, opErr = pkg.EvaluateMultiplication(lval, rval)
 		case OpDiv:
-			val, opErr = EvaluateDivision(lval, rval)
+			val, opErr = pkg.EvaluateDivision(lval, rval)
 		case OpMod:
-			val, opErr = EvaluateModulo(lval, rval)
+			val, opErr = pkg.EvaluateModulo(lval, rval)
 		case OpAdd:
-			val, opErr = EvaluateAddition(lval, rval)
+			val, opErr = pkg.EvaluateAddition(lval, rval)
 		case OpSub:
-			val, opErr = EvaluateSubstraction(lval, rval)
+			val, opErr = pkg.EvaluateSubstraction(lval, rval)
 		case OpBitAnd:
-			val, opErr = EvaluateBitAnd(lval, rval)
+			val, opErr = pkg.EvaluateBitAnd(lval, rval)
 		case OpBitOr:
-			val, opErr = EvaluateBitOr(lval, rval)
+			val, opErr = pkg.EvaluateBitOr(lval, rval)
 		case OpGT:
-			val, opErr = EvaluateGreaterThan(lval, rval)
+			val, opErr = pkg.EvaluateGreaterThan(lval, rval)
 		case OpLT:
-			val, opErr = EvaluateLesserThan(lval, rval)
+			val, opErr = pkg.EvaluateLesserThan(lval, rval)
 		case OpGTE:
-			val, opErr = EvaluateGreaterThanEqual(lval, rval)
+			val, opErr = pkg.EvaluateGreaterThanEqual(lval, rval)
 		case OpLTE:
-			val, opErr = EvaluateLesserThanEqual(lval, rval)
+			val, opErr = pkg.EvaluateLesserThanEqual(lval, rval)
 		case OpEq:
-			val, opErr = EvaluateEqual(lval, rval)
+			val, opErr = pkg.EvaluateEqual(lval, rval)
 		case OpNEq:
-			val, opErr = EvaluateNotEqual(lval, rval)
+			val, opErr = pkg.EvaluateNotEqual(lval, rval)
 		case OpAnd:
-			val, opErr = EvaluateLogicAnd(lval, rval)
+			val, opErr = pkg.EvaluateLogicAnd(lval, rval)
 		case OpOr:
-			val, opErr = EvaluateLogicOr(lval, rval)
+			val, opErr = pkg.EvaluateLogicOr(lval, rval)
 		}
 		if opErr == nil {
 			e.Value = val
@@ -240,742 +241,4 @@ func (e *Expression) Evaluate() (reflect.Value, error) {
 		return val, opErr
 	}
 	return reflect.ValueOf(nil), nil
-}
-
-// EvaluateMultiplication will evaluate multiplication operation over two value
-func EvaluateMultiplication(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv * rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv * int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) * rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not multipy data type of %s", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) * rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv * rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) * rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not multipy data type of %s", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv * float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv * float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv * rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not multipy data type of %s", right.Kind().String())
-		}
-	default:
-		return reflect.ValueOf(nil), fmt.Errorf("can not multipy data type of %s", left.Kind().String())
-	}
-}
-
-// EvaluateDivision will evaluate division operation over two value
-func EvaluateDivision(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(float64(lv) / float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(float64(lv) / float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) / rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in division", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(float64(lv) / float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(float64(lv) / float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) / rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in division", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv / float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv / float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv / rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in division", right.Kind().String())
-		}
-	default:
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in division", left.Kind().String())
-	}
-}
-
-// EvaluateModulo will evaluate modulo operation over two value
-func EvaluateModulo(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv % rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv % int64(rv)), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in modulo", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) % rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(int64(lv) % int64(rv)), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in modulo", right.Kind().String())
-		}
-	default:
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in modulo", left.Kind().String())
-	}
-}
-
-// EvaluateAddition will evaluate addition operation over two value
-func EvaluateAddition(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.String:
-		lv := left.String()
-		switch right.Kind() {
-		case reflect.String:
-			rv := right.String()
-			return reflect.ValueOf(fmt.Sprintf("%s%s", lv, rv)), nil
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(fmt.Sprintf("%s%d", lv, rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(fmt.Sprintf("%s%d", lv, rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(fmt.Sprintf("%s%f", lv, rv)), nil
-		case reflect.Bool:
-			rv := right.Bool()
-			return reflect.ValueOf(fmt.Sprintf("%s%v", lv, rv)), nil
-		default:
-			if right.Type().String() == "time.Time" {
-				rv := right.Interface().(time.Time)
-				return reflect.ValueOf(fmt.Sprintf("%s%s", lv, rv.Format(time.RFC3339))), nil
-			}
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in addition", right.Kind().String())
-		}
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.String:
-			rv := right.String()
-			return reflect.ValueOf(fmt.Sprintf("%d%s", lv, rv)), nil
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv + rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv + int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) + rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in addition", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.String:
-			rv := right.String()
-			return reflect.ValueOf(fmt.Sprintf("%d%s", lv, rv)), nil
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) + rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv + rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) + rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in division", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.String:
-			rv := right.String()
-			return reflect.ValueOf(fmt.Sprintf("%f%s", lv, rv)), nil
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv + float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv + float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv + rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in division", right.Kind().String())
-		}
-	default:
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in division", left.Kind().String())
-	}
-}
-
-// EvaluateSubstraction will evaluate substraction operation over two value
-func EvaluateSubstraction(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv - rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv - int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) - rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in substraction", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) - rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv - rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) - rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in substraction", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv - float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv - float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv - rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in substraction", right.Kind().String())
-		}
-	default:
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in substraction", left.Kind().String())
-	}
-}
-
-// EvaluateBitAnd will evaluate Bitwise And operation over two value
-func EvaluateBitAnd(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv & rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv & int64(rv)), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in bitwise AND operation", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) & rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv & rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in bitwise AND operation", right.Kind().String())
-		}
-	default:
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in bitwise AND operation", left.Kind().String())
-	}
-}
-
-// EvaluateBitOr will evaluate Bitwise Or operation over two value
-func EvaluateBitOr(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv | rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv | int64(rv)), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in bitwise OR operation", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) | rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv | rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in bitwise OR operation", right.Kind().String())
-		}
-	default:
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in bitwise OR operation", left.Kind().String())
-	}
-}
-
-// EvaluateGreaterThan will evaluate GreaterThan operation over two value
-func EvaluateGreaterThan(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv > rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv > int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) > rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GT comparison", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) > rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv > rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) > rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GT comparison", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv > float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv > float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv > rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GT comparison", right.Kind().String())
-		}
-	default:
-		if left.Type().String() == "time.Time" && right.Type().String() == "time.Time" {
-			lv := left.Interface().(time.Time)
-			rv := right.Interface().(time.Time)
-			return reflect.ValueOf(lv.After(rv)), nil
-		}
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GT comparison", left.Kind().String())
-	}
-}
-
-// EvaluateLesserThan will evaluate LesserThan operation over two value
-func EvaluateLesserThan(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv < rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv < int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) < rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LT comparison", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) < rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv < rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) < rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LT comparison", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv < float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv < float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv < rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LT comparison", right.Kind().String())
-		}
-	default:
-		if left.Type().String() == "time.Time" && right.Type().String() == "time.Time" {
-			lv := left.Interface().(time.Time)
-			rv := right.Interface().(time.Time)
-			return reflect.ValueOf(lv.Before(rv)), nil
-		}
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LT comparison", left.Kind().String())
-	}
-}
-
-// EvaluateGreaterThanEqual will evaluate GreaterThanEqual operation over two value
-func EvaluateGreaterThanEqual(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv >= rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv >= int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) >= rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GTE comparison", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) >= rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv >= rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) >= rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GTE comparison", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv >= float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv >= float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv >= rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GTE comparison", right.Kind().String())
-		}
-	default:
-		if left.Type().String() == "time.Time" && right.Type().String() == "time.Time" {
-			lv := left.Interface().(time.Time)
-			rv := right.Interface().(time.Time)
-			return reflect.ValueOf(lv.After(rv) || lv == rv), nil
-		}
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in GTE comparison", left.Kind().String())
-	}
-}
-
-// EvaluateLesserThanEqual will evaluate LesserThanEqual operation over two value
-func EvaluateLesserThanEqual(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv <= rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv <= int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) <= rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LTE comparison", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) <= rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv <= rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) <= rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LTE comparison", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv <= float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv <= float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv <= rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LTE comparison", right.Kind().String())
-		}
-	default:
-		if left.Type().String() == "time.Time" && right.Type().String() == "time.Time" {
-			lv := left.Interface().(time.Time)
-			rv := right.Interface().(time.Time)
-			return reflect.ValueOf(lv.Before(rv) || lv == rv), nil
-		}
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in LTE comparison", left.Kind().String())
-	}
-}
-
-// EvaluateEqual will evaluate Equal operation over two value
-func EvaluateEqual(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.String:
-		lv := left.String()
-		if right.Kind() == reflect.String {
-			rv := right.String()
-			return reflect.ValueOf(lv == rv), nil
-		}
-		return reflect.ValueOf(false), nil
-	case reflect.Bool:
-		lv := left.Bool()
-		if right.Kind() == reflect.Bool {
-			rv := right.Bool()
-			return reflect.ValueOf(lv == rv), nil
-		}
-		return reflect.ValueOf(false), nil
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv == rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv == int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) == rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) == rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv == rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) == rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv == float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv == float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv == rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", right.Kind().String())
-		}
-	default:
-		if left.Type().String() == "time.Time" && right.Type().String() == "time.Time" {
-			lv := left.Interface().(time.Time)
-			rv := right.Interface().(time.Time)
-			return reflect.ValueOf(lv == rv), nil
-		}
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", left.Kind().String())
-	}
-}
-
-// EvaluateNotEqual will evaluate NotEqual operation over two value
-func EvaluateNotEqual(left, right reflect.Value) (reflect.Value, error) {
-	switch left.Kind() {
-	case reflect.String:
-		lv := left.String()
-		if right.Kind() == reflect.String {
-			rv := right.String()
-			return reflect.ValueOf(lv != rv), nil
-		}
-		return reflect.ValueOf(false), nil
-	case reflect.Bool:
-		lv := left.Bool()
-		if right.Kind() == reflect.Bool {
-			rv := right.Bool()
-			return reflect.ValueOf(lv != rv), nil
-		}
-		return reflect.ValueOf(false), nil
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		lv := left.Int()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv != rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv != int64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) != rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", right.Kind().String())
-		}
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		lv := left.Uint()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(int64(lv) != rv), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv != rv), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(float64(lv) != rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", right.Kind().String())
-		}
-	case reflect.Float32, reflect.Float64:
-		lv := left.Float()
-		switch right.Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			rv := right.Int()
-			return reflect.ValueOf(lv != float64(rv)), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			rv := right.Uint()
-			return reflect.ValueOf(lv != float64(rv)), nil
-		case reflect.Float32, reflect.Float64:
-			rv := right.Float()
-			return reflect.ValueOf(lv != rv), nil
-		default:
-			return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", right.Kind().String())
-		}
-	default:
-		if left.Type().String() == "time.Time" && right.Type().String() == "time.Time" {
-			lv := left.Interface().(time.Time)
-			rv := right.Interface().(time.Time)
-			return reflect.ValueOf(lv != rv), nil
-		}
-		return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in EQ comparison", left.Kind().String())
-	}
-}
-
-// EvaluateLogicAnd will evaluate LogicalAnd operation over two value
-func EvaluateLogicAnd(left, right reflect.Value) (reflect.Value, error) {
-	if left.Kind() == reflect.Bool && right.Kind() == reflect.Bool {
-		lv := left.Bool()
-		rv := right.Bool()
-		return reflect.ValueOf(lv && rv), nil
-	}
-	return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in Logical AND comparison", left.Kind().String())
-}
-
-// EvaluateLogicOr will evaluate LogicalOr operation over two value
-func EvaluateLogicOr(left, right reflect.Value) (reflect.Value, error) {
-	if left.Kind() == reflect.Bool && right.Kind() == reflect.Bool {
-		lv := left.Bool()
-		rv := right.Bool()
-		return reflect.ValueOf(lv || rv), nil
-	}
-	return reflect.ValueOf(nil), fmt.Errorf("can not use data type of %s in Logical OR comparison", left.Kind().String())
 }
